Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   lib/prune.py

+++ lib/prune.py
@@ -130,8 +130,8 @@ def prune_magnitude(args, model, tokenizer, device=torch.device("cuda:0"), prune
             r, k = 16, 16
             # Given matrices
             w = W
-            a = torch.rand(r, n).to('cuda:0')
-            b = torch.rand(k, n).to('cuda:0')
+            a = torch.rand(r, n, dtype = w.dtype).to('cuda:0')
+            b = torch.rand(k, n, dtype = w.dtype).to('cuda:0')
 
             # Concatenate a and b horizontally
             c = torch.cat((a, b), dim=0).to('cuda:0')
